+++ R/combine_sources.R
@@ -18,15 +18,13 @@
 #' \code{\link{simmr_mcmc}}.
 #' @param to_combine The names of exactly two sources. These should match the
 #' names given to \code{\link{simmr_load}}.
-#' @param new_source_name A name to give to the new combined source. 
+#' @param new_source_name A name to give to the new combined source.
 #' @return A new \code{simmr_output} object
 #' @author Andrew Parnell <andrew.parnell@@mu.ie>
 #' @seealso See \code{\link{simmr_mcmc}} and the associated vignette for
 #' examples.
 #' @examples
 #' \dontrun{
-#' # Data set 1: 10 obs on 2 isos, 4 sources, with tefs and concdep
-#'
 #' # The data
 #' data(geese_data)
 #'
@@ -71,8 +69,8 @@
 #' plot(simmr_1_out, type = "matrix")
 #'
 #' simmr_out_combine <- combine_sources(simmr_1_out,
-#'   to_combine = c("Source A", "Source D"),
-#'   new_source_name = "Source A and D"
+#'   to_combine = c("U.lactuca", "Enteromorpha"),
+#'   new_source_name = "U.lac+Ent"
 #' )
 #' plot(simmr_out_combine$input)
 #' plot(simmr_out_combine, type = "boxplot", title = "simmr output: combined sources")
@@ -102,7 +100,7 @@ combine_sources.simmr_output <- function(simmr_out,
 
   # Find which columns to combine by number
   to_combine_cols <- sort(match(to_combine, simmr_out$input$source_names))
-  
+
   # Create a new object
   simmr_new_out <- simmr_out